--- src/components/NewProject/TextInput.js
@@ -1,17 +1,0 @@
-import React from "react";
-import { TextField } from "@material-ui/core";
-
-export default function TextInput({ name, label, autoFocus }) {
-  return (
-    <TextField
-      variant="outlined"
-      margin="normal"
-      required
-      fullWidth
-      id={name}
-      label={label}
-      name={name}
-      autoFocus={autoFocus ? true : false}
-    />
-  );
-}
