--- src/components/NewProject/createapi/Card.js
@@ -1,44 +1,0 @@
-import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import {
-  Card,
-  CardActions,
-  CardContent,
-  Button,
-  Typography,
-} from "@material-ui/core";
-
-const useStyles = makeStyles({
-  title: {
-    fontSize: 14,
-  },
-  pos: {
-    marginBottom: 12,
-  },
-});
-
-export default function SimpleCard({
-  title,
-  subtitle,
-  description,
-  linkText,
-  link,
-}) {
-  const classes = useStyles();
-
-  return (
-    <Card className={classes.root}>
-      <CardContent>
-        <Typography variant="h5" component="h2">
-          {title}
-        </Typography>
-        <Typography className={classes.pos} color="textSecondary">
-          {subtitle}
-        </Typography>
-        <Typography variant="body2" component="p">
-          {description}
-        </Typography>
-      </CardContent>
-    </Card>
-  );
-}
