+++ src/pages/home.js
@@ -1,37 +1,15 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import { Grid } from "@material-ui/core";
-import Card from "../components/Home/Card";
-import CardAdd from "../components/Home/CardAdd";
+import { useSelector } from "react-redux";
+import Home from "../components/Home";
 
-import { projects } from "./content/home";
+export default function HomePage() {
+  const projects = useSelector((state) => state.projects);
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  paper: {
-    padding: theme.spacing(2),
-    textAlign: "center",
-    color: theme.palette.text.secondary,
-  },
-}));
-
-export default function Home() {
-  const classes = useStyles();
+  if (!projects && projects.length <= 0) return null;
 
   return (
-    <div className={classes.root}>
-      <Grid container spacing={3}>
-        {projects.map((item, index) => (
-          <Grid item xs={12} sm={6} md={3} lg={4} key={item.title}>
-            <Card {...item} />
-          </Grid>
-        ))}
-        <Grid item xs={12} sm={6} md={3} lg={4}>
-          <CardAdd />
-        </Grid>
-      </Grid>
-    </div>
+    <>
+      <Home projects={projects} />
+    </>
   );
 }
